+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 
 * Relation archetypes are removed when they are empty *and* the target entity is dead (#238, #242)
 * Support an unlimited number of cached filters, instead of 128 (#245)
+* `WorldStats` contains the number of cached filters (#247)
+* Archetypes with entity relations are removed on `World.Reset` (#247)
 
 ### Performance